Fix for issues reported by Mahesh (MasterProgrammer) on the forums. There still are problems with comma separators which are not followed by a space. E.g.
	MOVE 10 to WS-A,WS-B
	DISPLAY A,B 

The problem is that the tokenizers can not decide whether or not the comma is a separator or not. So they don't tag it as a separator. This means that these commas get through to the parser, which then fails to match them.

The solution for now is to update the grammar to accept these commas where they make sense. I already updated the grammar rules for the MOVE and DISPLAY statements to fix the above examples. Extra tests are also included.

A more general solution still eludes me...

PS. I also modified some classes to make for easier debugging. These are not part of the fix, which is limited to Cobol.kg.

git-svn-id: svn+ssh://svn.code.sf.net/p/koopa/code/trunk@86 07b9d407-957a-469a-9197-ae794bbef2a3