Major contribution from Adrian Noguero:
"For the current version I needed to be able to parse conditions, so I tried to develop a new grammar for conditions that is gentle to the current tests. Moreover I made some improvements to the current implementation of the nestedStatemens that I think that provide better support for verbs without a finished implementation."

There is more here than I can describe. Suffice it to say that coverage of the testsuite has gone up, and is now at 98.92%! Adrian: thank you!

git-svn-id: svn+ssh://svn.code.sf.net/p/koopa/code/trunk@97 07b9d407-957a-469a-9197-ae794bbef2a3