Allowing integer constants wherever integers are allowed, and alpanumeric constants wherever alphanumerics are allowed.
Of course, Koopa can't verify whether the names being referenced as constants actually are constants. It assumes the best-case scenario.

Allowing this required some tweaking of the lowest-level grammar rules, which exposed some regressions due to ordering of choices in other grammar rules. I fixed all these, and all tests pass again. In fact, coverage of the testsuite went up as a result, and the number of warnings went down.

git-svn-id: svn+ssh://svn.code.sf.net/p/koopa/code/trunk@112 07b9d407-957a-469a-9197-ae794bbef2a3