Support for ">>" compiler directives.
Fix for issue #11 "not all input was consumed: compiler directives >> should be ignored", by Simon Sobisch.

Rework of the previous fix, this time based on info from the FCD (Final Committee Draft) of the next COBOL standard.
The actual compiler directive is not being checked; I'm only looking for the start marker.
The line gets marked as being a compiler directive, and now has a separate styling in the GUI.

git-svn-id: svn+ssh://svn.code.sf.net/p/koopa/code/trunk@130 07b9d407-957a-469a-9197-ae794bbef2a3